requirements.txt
streamlit==1.28.1
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
pandas==2.0.3
